+++ debug/prop.php
@@ -1,0 +1,8 @@
+<?php
+
+require __DIR__ . '/../vendor/autoload.php';
+
+$rnd = new \Random\Randomizer();
+
+var_dump($rnd->engine);
+var_dump($rnd->something);