+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Randomizer`
   * `getBytes($length)`
+  * `shuffleArray($array)`
 * `Mt19937`
   * Serialization and unserialization are now compatible with PHP 8.2
     if performed in PHP 7.4+