+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * `getBytes($length)`
   * `shuffleArray($array)`
   * `shuffleBytes($bytes)`
+  * `arrayPickKeys($array, $num)`
 * `Mt19937`
   * Serialization and unserialization are now compatible with PHP 8.2
     if performed in PHP 7.4+