+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.next
+## 0.2.0
+
+*Jul 27, 2022*
+
+NOTE: 0.2.0 currently is not fully compatible when using custom engines.
+This will be fixed in 0.3.0 after PHP 8.2.0 beta 2 is released with some critical fixes. 
 
 * `Randomizer`
   * `$engine`
@@ -8,6 +13,8 @@
   * `shuffleArray($array)`
   * `shuffleBytes($bytes)`
   * `arrayPickKeys($array, $num)`
+  * Serialization and unserialization are now compatible with PHP 8.2
+    if performed in PHP 7.4+
 * `Mt19937`
   * Serialization and unserialization are now compatible with PHP 8.2
     if performed in PHP 7.4+